separate culling and transform step
