+++ cocos2d/core/platform/CCMacro.js
@@ -1885,6 +1885,21 @@ cc.macro = {
      */
     ENABLE_GC_FOR_NATIVE_OBJECTS: true,
 
+
+    /**
+     * !#en
+     * Whether or not enable auto culling.
+     * If your game have more dynamic objects, we suggest to disable auto culling.
+     * If your game have more static objects, we suggest to enable auto culling.
+     * !#zh
+     * 是否开启自动裁减功能，开启裁减功能将会把在屏幕外的物体从渲染队列中去除掉。
+     * 如果游戏中的动态物体比较多的话，建议将此选项关闭。
+     * 如果游戏中的静态物体比较多的话，建议将此选项打开。
+     * @property {Boolean} ENABLE_CULLING
+     * @default true
+     */
+    ENABLE_CULLING: true,
+
     /**
      * !#en 
      * Whether or not enabled tiled map auto culling.
